refactor(tracker): split GitHub rate-limit classification into its own file
The headerless secondary-limit body probe pushed github.go to 802
physical lines, two over the machine-enforced 800-line production file
budget (TestProductionGoFilesStayWithinSizeBudget). Move the cohesive
rate-limit classification unit (githubRateLimited +
githubSecondaryLimitBody) to github_ratelimit.go — a by-responsibility
decomposition per AGENTS.md rule 7, not a baseline exemption.

Refs #761
